requirements.txt
streamlit
tensorflow
numpy
Pillow
streamlit-drawable-canvas
pandas
opencv-python